refactor: use sets to optimize further
by switching to sets from the objects we can take advantage of the
constraint that we are only concerned with maximum of pairs.

we can also use a for/of loop to avoid an extra o(n) operation to turn
the string into an array.

we don't need the conditional at the beginning for the "type" because
if there's a character remaining in the set, it _implies_ only one "odd"
occurrence.