feat: inplace shuffle
this approach uses the "fisher-yates shuffle" (also known as the "knuth
shuffle").

there are two big points here:
1. the range of the elements we can swap with are only what's "remaining"
that's because the inplace part of this is achieved by using the current
index as location for the swapped value
2. our getRandom takes a floor and a ceiling - which is what's remaining
of the array, and so finds a value from the array based on the available
indices.