fix: handle internal tree
the original implementation only conducted a check at the node/children
level.

this new approach inverts the level of interest to the node's ancestors.

by looking above the node, I can confirm that each node's value is
within the correct range, something not possible to do when lookign
down in a bst.