feat: memoized fibonacci
another attempt at this - playing around with memoization -- point is
that a single top level "general" memoization only helps with that
argument. if the function is recursive, all "internal" calculations are
still required.

the only way to fix that is to pass along the cache and modify the func
itself.