feat: improved fastestPath
by taking advantage of the fact that the maximum horizontal or vertical
distance will be the longest time, we can use a simpler calculation by
leveraging the the absolute value to find the distance.

now, instead of incremental steps, we can go directly to the time it
takes to traverse the distance.