✨ kLengthApart
O(n) time
O(1) space

done with both a while loop and for loop.

For loop is actually simpler here because of the check for undefined.

If the lists were not binary, it would have been less complicated,
but JS evaluating a 0 to false meant the check had to be more
verbose.